requirements.txt
transformers
torch
torchaudio
sentencepiece
soundfile
boto3
requests
spacy
websockets
python-dotenv